YOLOv4/YOLOv4_tensorflow/requirements.txt
tensorflow-gpu==1.14.0
numpy==1.18.1		
opencv-python==4.1.2.30
